Move yubico functions to a new file
